+++ NEWS.md
@@ -4,7 +4,9 @@ This page describes the most important changes in `TypeUtils`. The format is bas
 [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to
 [Semantic Versioning](https://semver.org).
 
-## Unreleased
+## Version 1.14.0 (2026-01-06)
+
+This new minor version essentially provides `lazymap` (see below).
 
 ### Added